Update higherlower.py
removed the following line "data.remove(B)" from the final else block of the while loop. In this version, no entries are every removed from data. While this means duplicate match ups are possible and inevitable on a large enough timescale, it also ensures the highest possible number of combinations of entries.